+++ packages/expo-updates/android/src/main/java/expo/modules/updates/manifest/ManifestFactory.kt
@@ -2,7 +2,6 @@ package expo.modules.updates.manifest
 
 import expo.modules.updates.UpdatesConfiguration
 import expo.modules.manifests.core.BareManifest
-import expo.modules.manifests.core.LegacyManifest
 import expo.modules.manifests.core.NewManifest
 import org.json.JSONException
 import org.json.JSONObject
@@ -16,8 +15,10 @@ object ManifestFactory {
   @Throws(Exception::class)
   fun getManifest(manifestJson: JSONObject, responseHeaderData: ResponseHeaderData, extensions: JSONObject?, configuration: UpdatesConfiguration): UpdateManifest {
     return when (val expoProtocolVersion = responseHeaderData.protocolVersion) {
+      // TODO(wschurman): remove error in a few major releases after SDK 51 when it's unlikely classic updates
+      // may erroneously be served
       null -> {
-        LegacyUpdateManifest.fromLegacyManifest(LegacyManifest(manifestJson), configuration)
+        throw Exception("Legacy manifests are no longer supported")
       }
       0, 1 -> {
         NewUpdateManifest.fromNewManifest(NewManifest(manifestJson), extensions, configuration)